user_authentication/requirements.txt
pip install django


django-admin startproject iblog


python manage.py startapp home

configure your app in settings.py in installed apps

create urls.py in app
create forms.py in app


In forms.py  - import
from django import forms
from django.contrib.auth.models import User
from django.contrib.auth.forms import UserCreationForm, PasswordResetForm

and make UserCreationForm


In views.py import
from .forms import SignUpForm
from django.contrib.auth import authenticate, login, logout

In urls.py import 
from . import views